pyproject.toml
[project]
requires-python = ">=3.12"
name = "openalgoUI"
version = "2.0.0.2"
description = "Broker-agnostic open-source trading automation"
dependencies = [
  "aniso8601==9.0.1",
  "annotated-types==0.7.0",
  "anyio==4.12.1",
  "appnope==0.1.4",
  "APScheduler==3.11.2",
  "argon2-cffi==23.1.0",
  "argon2-cffi-bindings==21.2.0",
  "asttokens==3.0.1",
  "attrs==24.2.0",
  "bcrypt==5.0.0",
  "bidict==0.23.1",
  "blinker==1.9.0",
  "cachetools==7.0.4",
  "certifi==2024.7.4",
  "cffi==2.0.0",
  "charset-normalizer==3.4.6",
  "choreographer==1.2.1",
  "click==8.3.1",
  "colorama==0.4.6",
  "comm==0.2.3",
  "cryptography==46.0.6",
  "darkdetect==0.8.0",
  "debugpy==1.8.20",
  "decorator==5.2.1",
  "Deprecated==1.3.1",
  "dnspython==2.8.0",
  "duckdb==1.5.0",
  "email-validator==2.3.0",
  "executing==2.2.1",
  "fastjsonschema==2.21.2",
  "Flask==3.1.3",
  "Flask-Bcrypt==1.0.1",
  "Flask-Cors==6.0.2",
  "Flask-Limiter==4.1.1",
  "Flask-Login==0.6.3",
  "flask-restx==1.3.2",
  "Flask-SocketIO==5.6.1",
  "Flask-SQLAlchemy==3.1.1",
  "Flask-WTF==1.2.2",
  "greenlet==3.3.2",
  "h11==0.16.0",
  "h2==4.3.0",
  "hpack==4.1.0",
  "httpcore==1.0.9",
  "httpx[http2]==0.28.1",
  "httpx-sse==0.4.3",
  "hyperframe==6.1.0",
  "idna==3.11",
  "importlib-resources==6.5.2",
  "iniconfig==2.3.0",
  "ipykernel==6.29.5",
  "ipython==9.11.0",
  "ipython-pygments-lexers==1.1.1",
  "itsdangerous==2.2.0",
  "jedi==0.19.2",
  "Jinja2==3.1.6",
  "jsonschema==4.26.0",
  "jsonschema-specifications==2023.12.1",
  "jupyter-client==8.8.0",
  "jupyter-core==5.9.1",
  "kaleido==1.2.0",
  "limits==5.8.0",
  "llvmlite==0.46.0b1",
  "logistro==2.0.1",
  "logzero==1.7.0",
  "markdown-it-py==4.0.0",
  "MarkupSafe==2.1.5",
  "marshmallow==3.26.2",
  "matplotlib-inline==0.2.1",
  "mcp==1.26.0",
  "mdurl==0.1.2",
  "py_lets_be_rational==1.0.1",
  "py_vollib==1.0.1",
  "narwhals==2.18.0",
  "nbformat==5.10.4",
  "nest-asyncio==1.6.0",
  "numba==0.64.0",
  "numpy>=2.0,<2.5",
  "openalgo==1.0.46",
  "ordered-set==4.1.0",
  "orjson==3.11.7",
  "packaging==24.1",
  "pandas==2.3.3",
  "pyarrow==23.0.1",
  "fastparquet==2025.12.0",
  "parso==0.8.6",
  "pexpect==4.9.0",
  "pillow>=12.1.1",
  "platformdirs==4.9.4",
  "plotly==6.6.0",
  "pluggy==1.6.0",
  "prompt-toolkit==3.0.52",
  "protobuf==6.33.5",
  "psutil==7.2.2",
  "ptyprocess==0.7.0",
  "pure-eval==0.2.3",
  "pycparser==2.22",
  "pydantic==2.12.5",
  "pydantic-core==2.41.5",
  "pydantic-settings==2.13.1",
  "Pygments==2.20.0",
  "PyJWT==2.12.1",
  "pyngrok==7.5.1",
  "pyotp==2.9.0",
  "pytest==9.0.2",
  "pytest-timeout==2.4.0",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.2.2",
  "python-engineio==4.13.1",
  "python-multipart==0.0.22",
  "python-socketio==5.16.1",
  "python-telegram-bot==22.6",
  "pytz==2026.1.post1",
  "PyYAML==6.0.3",
  "pyzmq==27.1.0",
  "qrcode==8.2",
  "referencing==0.37.0",
  "requests==2.33.0",
  "rich==13.7.1",
  "rpds-py==0.30.0",
  "scipy==1.17.1",
  "setuptools==80.10.1",
  "simple-websocket==1.1.0",
  "simplejson==3.20.2",
  "six==1.17.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.48",
  "sse-starlette==2.4.1",
  "stack-data==0.6.3",
  "starlette==0.52.1",
  "tornado==6.5.5",
  "traitlets==5.14.3",
  "typing-extensions>=4.12.2",
  "typing-inspection>=0.4.1",
  "tzdata==2025.3",
  "tzlocal==5.3.1",
  "urllib3==2.6.3",
  "uvicorn==0.42.0",
  "wcwidth==0.6.0",
  "websocket-client==1.9.0",
  "websockets==15.0.1",
  "Werkzeug==3.1.6",
  "wheel==0.46.3",
  "wrapt==1.16.0",
  "wsproto==1.3.2",
  "WTForms==3.2.1",
  "zipp==3.23.0",
  "zmq==0.0.0",
  "loguru>=0.7.3",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "telethon>=1.42.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

# Ruff - Fast Python linter and formatter
[tool.ruff]
line-length = 100
target-version = "py312"
exclude = [
    ".venv",
    "frontend",
    "node_modules",
    "__pycache__",
    "*.pyc",
    "db",
    "log",
    "strategies",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "E402",   # module level import not at top of file
    "F401",   # imported but unused (sometimes intentional)
]

[tool.ruff.lint.isort]
known-first-party = ["broker", "blueprints", "database", "services", "utils", "restx_api", "extensions", "limiter", "cors", "csp"]

# This is not a distributable Python package — tell setuptools not to discover packages
[tool.setuptools]
packages = []

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --timeout=60"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[dependency-groups]
dev = [
    "bandit[sarif]>=1.9.3",
    "detect-secrets>=1.5.0",
    "pip-audit>=2.10.0",
    "pytest-timeout>=2.4.0",
    "ruff>=0.14.14",
    "vulture>=2.15",
]
